requirements.txt
python-telegram-bot==20.3
Pillow>=10.0.0
pytesseract==0.3.10
numpy>=1.24.3
python-dateutil==2.8.2
SQLAlchemy>=2.0.15
matplotlib>=3.7.1
python-dotenv==1.0.0
pandas>=2.0.0
tabulate==0.9.0 